+++ tests/Fixtures/inc/Engine/Optimization/CSS/CriticalPath/CriticalCSSSubscriber/generateCriticalCssOnActivation.php
